将上传附件功能迁移到AttachmentService中...

+++ info-core/info-dms/src/main/java/com/vertonur/dms/exception/AttachmentSizeExceedException.java
@@ -5,7 +5,7 @@ public class AttachmentSizeExceedException extends Exception {
 
 	private static final long serialVersionUID = -2305309245986580825L;
 
-	public AttachmentSizeExceedException(int fileSize, int configuredSize) {
+	public AttachmentSizeExceedException(long fileSize, long configuredSize) {
 		super(
 				"Size of the uploaded attachment exceeds the configured size, attachment size:"
 						+ fileSize + ",configured size:" + configuredSize);
